fix():
The ProcessorController doesn't have a validate filter for the delay—there's no explicit 400 response for an invalid delay.

The cookieValidationKey in the committed params.php is a secret in the repository; the files are listed in .gitignore, but they're in git.

RequestController injects RequestService but doesn't use it.

The amount in the specification is an integer, and the rule in the model is number.

The rule for status in Request::rules() is commented out.

+++ migrations/m260316_165804_create_table_requests.php
@@ -11,7 +11,7 @@ class m260316_165804_create_table_requests extends Migration
         $this->createTable($this->tableName, [
             'id' => $this->primaryKey()->comment('идентификатор заявки.'),
             'user_id' => $this->integer()->notNull()->comment('идентификатор пользователя, подающего заявку.'),
-            'amount' => $this->decimal(10, 2)->notNull()->comment('сумма займа, которую пользователь запрашивает.'),
+            'amount' => $this->integer()->notNull()->comment('сумма займа, которую пользователь запрашивает.'),
             'term' => $this->integer()->notNull()->comment('срок займа в днях'),
             'status' => $this->integer()->notNull()->defaultValue(Status::WAIT->value)->comment('статус заявки (например, "wait", "approved", "declined").'),
         ]);
